Move ValidateResourceExist(s) from pkg/service to pkg/devfile/adapters/kubernetes/component
